% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transform.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Cluster Transformation Matrix}
\usage{
cluster_matrix(cluster)
}
\arguments{
\item{cluster}{A vector of cluster ids.}
}
\value{
A numeric matrix.
}
\description{
Direct translation of \code{cluster_matrix.mata}.
}
\keyword{internal}
