% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean_sigma.R
\name{demean_sigma}
\alias{demean_sigma}
\title{Demean a Covariance Matrix}
\usage{
demean_sigma(sigma)
}
\arguments{
\item{sigma}{A numeric covariance matrix.}
}
\value{
A numeric matrix that has been demeaned.
}
\description{
Given a covariance matrix \code{sigma}, this function computes the variance of
\eqn{X - \bar{X}}, effectively demeaning the matrix by removing the influence of the mean.
}
