% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cbar.R
\name{get_cbar}
\alias{get_cbar}
\title{Compute cbar Value}
\usage{
get_cbar(rhobar, distmat)
}
\arguments{
\item{rhobar}{A numeric scalar representing the target value.}

\item{distmat}{A numeric matrix of distances.}
}
\value{
A numeric scalar \code{cbar} satisfying the desired property.
}
\description{
Determines the value of \code{cbar} such that \eqn{mean(exp(-cbar \times vd))} is approximately equal to
a given \code{rhobar}, where \code{vd} is the vectorized lower-triangular portion (including the diagonal)
of the distance matrix \code{distmat}. This function uses iterative adjustments and bisection to find
an appropriate value.
}
\keyword{internal}
