% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ha_param_i1_residual.R
\name{get_ha_param_i1_residual}
\alias{get_ha_param_i1_residual}
\title{Compute Ha Parameter for I(1) Residual Test}
\usage{
get_ha_param_i1_residual(om_ho, distmat, R, e, M)
}
\arguments{
\item{om_ho}{Null covariance matrix in low-frequency space.}

\item{distmat}{Distance matrix.}

\item{R}{Eigenvector matrix.}

\item{e}{Monte Carlo draw matrix.}

\item{M}{Residual-maker matrix.}
}
\value{
Scalar persistence parameter.
}
\description{
Translation of SPUR Mata \code{get_ha_parm_I1_residual.mata}.
}
\keyword{internal}
