% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_helpers_covariance.R
\name{get_sigma_dm}
\alias{get_sigma_dm}
\title{Compute Demeaned Sigma (c)}
\usage{
get_sigma_dm(distmat, c)
}
\arguments{
\item{distmat}{A numeric matrix of distances.}

\item{c}{A numeric constant used in the exponential transformation.}
}
\value{
A numeric matrix representing the demeaned sigma.
}
\description{
Computes a demeaned version of \eqn{\sigma(c)}, where
\eqn{\sigma(c) = \exp(-c \times \text{distmat})}, and then demeans the resulting matrix.
}
\keyword{internal}
