% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_helpers_covariance.R
\name{get_sigma_lbm}
\alias{get_sigma_lbm}
\title{Compute LBM Covariance Matrix}
\usage{
get_sigma_lbm(distmat)
}
\arguments{
\item{distmat}{A numeric matrix of distances.}
}
\value{
A numeric matrix representing the LBM covariance matrix.
}
\description{
Computes the LBM covariance matrix from the given distance matrix, using the
first location as the origin.
}
\keyword{internal}
