% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma_residual.R
\name{get_sigma_residual}
\alias{get_sigma_residual}
\title{Residual Covariance Matrix}
\usage{
get_sigma_residual(distmat, c, M)
}
\arguments{
\item{distmat}{A distance matrix.}

\item{c}{Spatial persistence parameter.}

\item{M}{Residual-maker matrix.}
}
\value{
Residual covariance matrix.
}
\description{
Translation of SPUR Mata \code{get_sigma_residual.mata}.
}
\keyword{internal}
