% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transform.R
\name{lbm_gls_matrix}
\alias{lbm_gls_matrix}
\title{LBM-GLS Transformation Matrix}
\usage{
lbm_gls_matrix(s, latlong = FALSE)
}
\arguments{
\item{s}{A numeric matrix of spatial coordinates.}

\item{latlong}{Logical; if \code{TRUE}, coordinates are latitude/longitude.}
}
\value{
A numeric matrix.
}
\description{
Direct translation of \code{lbm_gls_matrix.mata}.
}
\keyword{internal}
