% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_helpers_matrix.R
\name{lvech}
\alias{lvech}
\title{Lower Triangular Vectorization}
\usage{
lvech(S)
}
\arguments{
\item{S}{A numeric matrix.}
}
\value{
A numeric vector containing the lower-triangular elements.
}
\description{
Extracts the elements of a matrix that are strictly below the main diagonal
and returns them as a vector.
}
\keyword{internal}
