% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transform.R
\name{nn_matrix}
\alias{nn_matrix}
\title{Nearest-Neighbor Transformation Matrix}
\usage{
nn_matrix(s, latlong = FALSE)
}
\arguments{
\item{s}{A numeric matrix of spatial coordinates.}

\item{latlong}{Logical; if \code{TRUE}, coordinates are latitude/longitude.}
}
\value{
A numeric matrix.
}
\description{
Direct translation of \code{nn_matrix.mata}.
}
\keyword{internal}
