% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_i1_test.R
\name{spur_i1_test}
\alias{spur_i1_test}
\title{Core SPUR I(1) Test Function}
\usage{
spur_i1_test(Y, distmat, emat)
}
\arguments{
\item{Y}{A numeric matrix (n x number_of_series).}

\item{distmat}{A numeric distance matrix (n x n).}

\item{emat}{A numeric simulation matrix (q x number_of_simulations).}
}
\value{
A list with elements: LR, pvalue, cv_vec, and ha_parm.
}
\description{
This function implements the spatial I(1) test. It is called internally by \code{spurtest_i1()}.
}
\keyword{internal}
