% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_persistence.R
\name{spur_persistence}
\alias{spur_persistence}
\title{Spatial Persistence Core Routine}
\usage{
spur_persistence(Z, distmat, emat, level)
}
\arguments{
\item{Z}{Outcome matrix.}

\item{distmat}{Normalized distance matrix.}

\item{emat}{Monte Carlo draw matrix.}

\item{level}{Confidence level between 0 and 1.}
}
\value{
List with lower/upper half-life confidence bounds.
}
\description{
Internal translation of SPUR Mata \code{spatial_persistence} and \code{c_ci}.
}
\keyword{internal}
