% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur_helpers_covariance.R
\name{get_sigma_lbm_dm}
\alias{get_sigma_lbm_dm}
\title{Compute Demeaned Sigma LBM}
\usage{
get_sigma_lbm_dm(distmat)
}
\arguments{
\item{distmat}{A numeric matrix of distances.}
}
\value{
A numeric matrix representing the demeaned sigma_lbm.
}
\description{
Computes the sigma_lbm matrix using the provided distance matrix and then
demeans it.
}
\keyword{internal}
